requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pandas==2.2.2
statsmodels==0.14.2
requests==2.31.0
numpy>=1.22.3
scipy!=1.9.2
python-dotenv